requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
oauth2client
html2image
pillow
Jinja2
pandas
discord.py
requests